Don't rely on size_t overflow

+++ include/string.hpp
@@ -40,8 +40,9 @@ public:
       end += size();
     auto searched_portion = substr(start, end - start);
     size_t seen = 0;
-    size_t pos = -1;
-    while ((pos = searched_portion.find(sub, pos + 1)) != string::npos) {
+    int pos = -1;
+    while ((pos = searched_portion.find(sub, static_cast<size_t>(pos + 1))) !=
+           string::npos) {
       seen++;
     }
     return seen;
